[fe] refactor: 로그인 코드 스플릿
