+++ client/src/Layout/Header.tsx
@@ -2,7 +2,7 @@
 
 import Link from "next/link";
 import { useState } from "react";
-import { useModalActions } from "../components/login";
+import { useModalActions } from "@/components/Login/contextAPI/useModal";
 
 /** 2023/05/04 - 헤더 컴포넌트 - by Kadesti */
 const Header: React.FC = () => {
