+++ client/src/Layout/index.tsx
@@ -8,13 +8,14 @@ import Nav from "./Nav";
 import Main from "./Main";
 import Footer from "./Footer";
 
-import { ModalProvider, Modal } from "../components/login";
+import ModalProvider from "@/components/Login/contextAPI/ModalProvider";
+import LoginModal from "../components/Login/LoginModal";
 
 /** 2023/05/10 - 화면 구성의 Root 경로 - by Kadesti */
 const Layout: React.FC<React.PropsWithChildren> = ({ children }) => (
   <LoadingProvider>
     <ModalProvider>
-      <Modal />
+      {/* { modalValue && <LoginModal />} */}
       <Header />
       <Nav />
       <Main>{children}</Main>
